backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
psycopg[binary]==3.2.1
pandas==2.2.2
numpy==1.26.4
pyjwt==2.9.0
bcrypt==4.2.0
pydantic==2.8.2
httpx==0.27.2
openai==1.40.6
python-dateutil==2.9.0.post0
